+++ packages/backend/src/server/api/endpoints/i/move.ts
@@ -43,11 +43,6 @@ export const meta = {
 			code: "NOT_REMOTE",
 			id: "4362f8dc-731f-4ad8-a694-be2a88922a24",
 		},
-		adminForbidden: {
-			message: "Admins cant migrate.",
-			code: "NOT_ADMIN_FORBIDDEN",
-			id: "4362e8dc-731f-4ad8-a694-be2a88922a24",
-		},
 		noSuchUser: {
 			message: "No such user.",
 			code: "NO_SUCH_USER",
@@ -93,7 +88,6 @@ function moveActivity(toUrl: string, fromUrl: string) {
 
 export default define(meta, paramDef, async (ps, user) => {
 	if (!ps.moveToAccount) throw new ApiError(meta.errors.noSuchMoveTarget);
-	if (user.isAdmin) throw new ApiError(meta.errors.adminForbidden);
 	if (user.movedToUri) throw new ApiError(meta.errors.alreadyMoved);
 
 	const { username, host } = parse(ps.moveToAccount);